+++ src/infrastructure/webgl/lib/vectorPositions.spec.ts
