+++ src/infrastructure/webgl/lib/vectorPositions.ts
