+++ src/infrastructure/webgl/lib/vertices.ts
