+++ src/infrastructure/webgl/programs/showComposition.frag
