+++ src/infrastructure/webgl/programs/showCompositionMap.frag
