+++ src/infrastructure/webgl/programs/showCompositionMap.vert
