+++ src/infrastructure/webgl/programs/showCompositionVectors.frag
