+++ src/infrastructure/webgl/programs/showCompositionVectors.vert
