+++ src/infrastructure/webgl/programs/showGrid.frag
