+++ src/infrastructure/webgl/programs/showLayerPoints.frag
