+++ src/infrastructure/webgl/programs/showLayerPoints.ts
