+++ src/infrastructure/webgl/programs/showLayerPoints.vert
