+++ src/infrastructure/webgl/programs/showTexture.frag
