+++ src/infrastructure/webgl/programs/showTexture.vert
