+++ src/infrastructure/webgl/programs/showTextureMap.frag
