+++ src/infrastructure/webgl/programs/showTextureMap.vert
