+++ src/infrastructure/webgl/programs/utils.ts
